add a function contains all th common code
