+++ lib/id/id-details.js
@@ -1,4 +1,3 @@
-const assignement = require('../shared/assignement');
 
 const nonNullIndexes = require('../shared/non-null-indexes');
 const trackDist = require('../shared/track-dist');
@@ -39,7 +38,6 @@ module.exports = function (options) {
 			throw (new Error(`Empty prediction track at index ${i}`));
 		}
 	});
-	const {vt, vc} = idShared.idDetails(groundTruths, predictions);
 
 	const maxCost = nFrames * 2;
 	const detailedDistFn = (gt, pred) => trackDist.detailed(gt, pred, options);
@@ -89,53 +87,19 @@ module.exports = function (options) {
 		return details;
 	};
 
-	const {matched, unmatched} = assignement(
-		vt,
-		vc,
-		{
-			threshold: maxCost * 2,
-			distFn: (a, b) => {
-				const {fp, fn} = binaryDistFn(a, b);
-				return fp + fn;
-			}
-		}
-	);
-
-	if (unmatched[0].length > 0 || unmatched[1].length > 0) {
-		console.log({unmatched, maxCost});
-		throw (new Error('ID metrics should not have unmatched assignement'));
-	}
-
-	// Console.log(matched.map(({indexes}) => [indexes, typeof(vt[indexes[0]].gtIndex) === 'number', typeof(vc[indexes[1]].predIndex) === 'number']))
-	const idtpSet = matched.filter(({indexes}) => typeof (vt[indexes[0]].gtIndex) === 'number' && typeof (vc[indexes[1]].predIndex) === 'number');
-
-	const match = idtpSet.map(({indexes}) => [vt[indexes[0]].gtIndex, vc[indexes[1]].predIndex]);
+	const distFn = (a, b) => {
+		const {fp, fn} = binaryDistFn(a, b);
+		return fp + fn;
+	};
 
 	const sum = array => array.reduce((a, b) => a + b, 0);
-
-	const gtMatch = [];
-	const predMatch = [];
-
-	matched.forEach(({indexes}) => {
-		if (typeof (vt[indexes[0]].gtIndex) === 'number') {
-			gtMatch[vt[indexes[0]].gtIndex] = vc[indexes[1]];
-		}
-
-		if (typeof (vc[indexes[1]].predIndex) === 'number') {
-			predMatch[vc[indexes[1]].predIndex] = vt[indexes[0]];
-		}
-	});
+	const {match, predMatch, gtMatch, tpPd: idtpIndexesPred, tpGT: idtpIndexesGt, fpIdx: idfpIndexes, fnIdx: idfnIndexes} = idShared.idDetails(groundTruths, predictions, {maxCost}, distFn, binaryDistFn);
 
 	const idfn = sum(groundTruths.map((_, gtIndex) => binaryDistFn({gtIndex}, gtMatch[gtIndex]).fn));
 	const idfp = sum(predictions.map((_, predIndex) => binaryDistFn(predMatch[predIndex], {predIndex}).fp));
 
-	const idfnIndexes = groundTruths.map((_, gtIndex) => binaryDistFn({gtIndex}, gtMatch[gtIndex]).fnIndexes);
-	const idfpIndexes = predictions.map((_, predIndex) => binaryDistFn(predMatch[predIndex], {predIndex}).fpIndexes);
-
 	const idtp1 = sum(groundTruths.map(p => p.filter(a => a !== null).length)) - idfn;
 	const idtp2 = sum(predictions.map(g => g.filter(a => a !== null).length)) - idfp;
-	const idtpIndexesGt = groundTruths.map((p, gtIndex) => nonNullIndexes(p).filter(index => !idfnIndexes[gtIndex].includes(index)));
-	const idtpIndexesPred = predictions.map((p, predIndex) => nonNullIndexes(p).filter(index => !idfpIndexes[predIndex].includes(index)));
 
 	if (idtp1 !== idtp2) {
 		throw (new Error('Should verify (8) : IDTP = Sum(τ∈AT, len(τ) − IDFN) = Sum(γ∈AC, len(γ) − IDFP)'));
