+++ lib/idcl/ideucl-details.js
@@ -1,4 +1,3 @@
-const assignement = require('../shared/assignement');
 const nonNullIndexes = require('../shared/non-null-indexes');
 const trackDist = require('../shared/track-dist');
 const idShared = require('../shared/id-shared');
@@ -88,7 +87,6 @@ module.exports = function (options) {
 			throw (new Error(`Empty prediction track at index ${i}`));
 		}
 	});
-	const {vt, vc} = idShared.idDetails(groundTruths, predictions);
 
 	const maxCost = ntraveld;
 	const detailedDistFn = (gt, pred) => trackDist.detailed(gt, pred, options);
@@ -146,48 +144,16 @@ module.exports = function (options) {
 		return details;
 	};
 
-	const {matched, unmatched} = assignement(
-		vt,
-		vc,
-		{
-			threshold: maxCost,
-			distFn: (a, b) => {
-				const dTravel = -binaryDistFn(a, b).dTravel;
-				return dTravel;
-			}
-		}
-
-	);
-
-	if (unmatched[0].length > 0 || unmatched[1].length > 0) {
-		console.log({unmatched, maxCost});
-		throw (new Error('ID metrics should not have unmatched assignement'));
-	}
+	const distFn = (a, b) => {
+		const dTravel = -binaryDistFn(a, b).dTravel;
+		return dTravel;
+	};
 
 	const sum = array => array.reduce((a, b) => a + b, 0);
-	const gtMatch = [];
-	const predMatch = [];
-
-	matched.forEach(({indexes}) => {
-		if (typeof (vt[indexes[0]].gtIndex) === 'number') {
-			gtMatch[vt[indexes[0]].gtIndex] = vc[indexes[1]];
-		}
-
-		if (typeof (vc[indexes[1]].predIndex) === 'number') {
-			predMatch[vc[indexes[1]].predIndex] = vt[indexes[0]];
-		}
-	});
-
-	const idEucltpSet = matched.filter(({indexes}) => typeof (vt[indexes[0]].gtIndex) === 'number' && typeof (vc[indexes[1]].predIndex) === 'number');
-
-	const match = idEucltpSet.map(({indexes}) => [vt[indexes[0]].gtIndex, vc[indexes[1]].predIndex]);
+	const {match, gtMatch, tpPd: idEucltpIndexesPred, tpGT: idEucltpIndexesGt, fpIdx: idEuclfpIndexes, fnIdx: idEuclfnIndexes} = idShared.idDetails(groundTruths, predictions, {maxCost}, distFn, binaryDistFn);
 
 	const dTravel = sum(groundTruths.map((_, gtIndex) => binaryDistFn({gtIndex}, gtMatch[gtIndex]).dTravel));
 	const trueD = sum(groundTruths.map((_, gtIndex) => distanceTraveled(groundTruths[gtIndex])));
-	const idEuclfnIndexes = groundTruths.map((_, gtIndex) => binaryDistFn({gtIndex}, gtMatch[gtIndex]).fnIndexes);
-	const idEuclfpIndexes = predictions.map((_, predIndex) => binaryDistFn(predMatch[predIndex], {predIndex}).fpIndexes);
-	const idEucltpIndexesGt = groundTruths.map((p, gtIndex) => nonNullIndexes(p).filter(index => !idEuclfnIndexes[gtIndex].includes(index)));
-	const idEucltpIndexesPred = predictions.map((p, predIndex) => nonNullIndexes(p).filter(index => !idEuclfpIndexes[predIndex].includes(index)));
 
 	const idEucl = dTravel / trueD;
 	const result = {idEucl, match, idEuclfpIndexes, idEuclfnIndexes, idEucltpIndexesPred, idEucltpIndexesGt};
